requirements.txt
streamlit[pdf]>=1.52.0
psycopg[binary]>=3.2.0
pandas>=2.2.0
numpy>=2.0.0
altair>=5.3.0
qrcode>=7.4.2
pillow>=10.4.0
reportlab>=4.2.0
openai>=1.56.0
requests>=2.32.0
python-jose[cryptography]>=3.3.0
